refactor: simplify trigger-async API route by moving logic to Temporal client
- Move scheduled_for handling from API route to NotificationClient
- API route now simply calls Temporal client without pre-fetching data
- Temporal client handles all scheduling logic internally
- Return type includes scheduledFor and startDelay for transparency
