+++ lib/temporal/client/notification-client.ts
@@ -2,11 +2,17 @@ import { v4 as uuidv4 } from 'uuid'
 import { getTemporalClient } from './index'
 import { notificationTriggerWorkflow, triggerMultipleNotificationsWorkflow } from '../workflows'
 import type { TriggerResult } from '@/lib/notifications/trigger'
+import { createClient } from '@supabase/supabase-js'
+import type { Database } from '@/lib/supabase/database.types'
+
+// Initialize Supabase client
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!
+const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
+const supabase = createClient<Database>(supabaseUrl, supabaseKey)
 
 export interface AsyncTriggerOptions {
   workflowId?: string
   taskQueue?: string
-  startDelay?: number | string
 }
 
 export class NotificationClient {
@@ -22,8 +28,30 @@ export class NotificationClient {
   async asyncTriggerNotificationById(
     notificationId: number,
     options?: AsyncTriggerOptions
-  ): Promise<{ workflowId: string; runId: string }> {
+  ): Promise<{ workflowId: string; runId: string; scheduledFor?: string; startDelay?: number }> {
     const client = await getTemporalClient()
+    
+    // Fetch notification to check scheduled_for
+    const { data: notification, error } = await supabase
+      .schema('notify')
+      .from('ent_notification')
+      .select('scheduled_for')
+      .eq('id', notificationId)
+      .single()
+
+    if (error) {
+      throw new Error(`Failed to fetch notification: ${error.message}`)
+    }
+
+    let startDelay: number | undefined
+    if (notification?.scheduled_for) {
+      const scheduledTime = new Date(notification.scheduled_for)
+      const now = new Date()
+      const delayMs = scheduledTime.getTime() - now.getTime()
+      if (delayMs > 0) {
+        startDelay = delayMs
+      }
+    }
     
     const workflowId = options?.workflowId || `trigger-notification-${notificationId}-${uuidv4()}`
     const taskQueue = options?.taskQueue || this.taskQueue
@@ -32,12 +60,14 @@ export class NotificationClient {
       args: [{ notificationId }],
       taskQueue,
       workflowId,
-      startDelay: options?.startDelay,
+      startDelay,
     })
 
     return {
       workflowId: handle.workflowId,
       runId: handle.firstExecutionRunId,
+      scheduledFor: notification?.scheduled_for || undefined,
+      startDelay,
     }
   }
 
@@ -57,7 +87,6 @@ export class NotificationClient {
       args: [{ notificationIds }],
       taskQueue,
       workflowId,
-      startDelay: options?.startDelay,
     })
 
     return {
